a bunch of changes:
using NPCCreatures plugin directly instead of converting bits of the code to scala.
added parser for longs
moved some Player pimping stuff into LivingEntity, so it can work on all living entities, not just players.
wolf-sheep is kind of working with the NetLogo plugin, except for a bug that sometimes crashes it on entity death. still working on it.
moved NPC to NPCTest, because it really is just for testing things out.