for better or worse, rolled some common commands into core.
this was mainly so that they could be used by the netlogo plugin. soon, ill refactor the other plugins to use them as well. one serious problem with this though, is that now they will overlap...what to do about that? just not have them installed at the same time? let them both handle the command? ugh.